door, canvas adjustments, game now almost fully playable
