+++ Assets/Saves/SaveData.json
@@ -1,1 +1,1 @@
-{"inventoryItems":[{"instanceID":26566},{"instanceID":26566},{"instanceID":26566},{"instanceID":26566},{"instanceID":26566},{"instanceID":26566},{"instanceID":26566},{"instanceID":26566}],"NPC_Pass":[true,true,true,true,true,true,true,true,true],"Items_Collected":[false,false,false,false,false,false,false,false,false],"tracker":[true,true,true,true,true,true,true,true,true],"levelTracker":[true,true,true]}
+{"inventoryItems":[{"instanceID":26572},{"instanceID":26572}],"NPC_Pass":[false,false,false,false,false,false,false,false,false],"Items_Collected":[false,false,false,false,false,false,false,false,false],"tracker":[true,true,true,true,true,true,true,true,true],"levelTracker":[true,true,true]}
