+++ Assets/SchoolRPG/ScriptableObjects/Inventory/Items/Cigarette.asset
@@ -17,5 +17,5 @@ MonoBehaviour:
   <Icon>k__BackingField: {fileID: 21300000, guid: 3c545196b77f68848be34b7bf72a0290, type: 3}
   <FullIcon>k__BackingField: {fileID: 21300000, guid: a1d4749c067a47c4182236bbd9505b98, type: 3}
   <Count>k__BackingField: 0
-  <Collected>k__BackingField: 1
+  <Collected>k__BackingField: 0
   <Id>k__BackingField: 6
