+++ Assets/SchoolRPG/ScriptableObjects/NPC/1.asset
@@ -15,12 +15,12 @@ MonoBehaviour:
   <Name>k__BackingField: Teacher
   <RequiredItem>k__BackingField: {fileID: 11400000, guid: 39aef5e2864a87d4dabc03a997e47337, type: 2}
   <Level>k__BackingField: 1
-  <IsPassed>k__BackingField: 1
+  <IsPassed>k__BackingField: 0
   Dialogue:
-  - 1
+  - dialogue
   PassDialogue:
-  - 2
+  - you passed
   AlreadyPassedDialogue:
-  - 3
+  - already passed
   NoPassDialogue:
-  - 4
+  - no pass
