+++ Assets/SchoolRPG/ScriptableObjects/NPC/2.asset
@@ -15,12 +15,12 @@ MonoBehaviour:
   <Name>k__BackingField: Claws lady
   <RequiredItem>k__BackingField: {fileID: 11400000, guid: 1eedea7aa89d499469aaf212bc815787, type: 2}
   <Level>k__BackingField: 1
-  <IsPassed>k__BackingField: 1
+  <IsPassed>k__BackingField: 0
   Dialogue:
-  - 1
+  - dialogue
   PassDialogue:
-  - 2
+  - you passed
   AlreadyPassedDialogue:
-  - 3
+  - already passed
   NoPassDialogue:
-  - 4
+  - no pass
