+++ Assets/SchoolRPG/ScriptableObjects/NPC/3.asset
@@ -15,12 +15,12 @@ MonoBehaviour:
   <Name>k__BackingField: 3
   <RequiredItem>k__BackingField: {fileID: 11400000, guid: 71e17667b5b6504488f1d30f37df6787, type: 2}
   <Level>k__BackingField: 2
-  <IsPassed>k__BackingField: 1
+  <IsPassed>k__BackingField: 0
   Dialogue:
-  - not passed
+  - dialogue
   PassDialogue:
   - you passed
   AlreadyPassedDialogue:
   - already passed
   NoPassDialogue:
-  - 4
+  - no pass
