+++ Assets/SchoolRPG/ScriptableObjects/NPC/4.asset
@@ -15,12 +15,12 @@ MonoBehaviour:
   <Name>k__BackingField: 4
   <RequiredItem>k__BackingField: {fileID: 11400000, guid: 8117705f93b698846823bcb104673e6e, type: 2}
   <Level>k__BackingField: 2
-  <IsPassed>k__BackingField: 1
+  <IsPassed>k__BackingField: 0
   Dialogue:
-  - not passed
+  - dialogue
   PassDialogue:
   - you passed
   AlreadyPassedDialogue:
   - already passed
   NoPassDialogue:
-  - 4
+  - no pass
