+++ Assets/SchoolRPG/ScriptableObjects/NPC/5.asset
@@ -15,12 +15,12 @@ MonoBehaviour:
   <Name>k__BackingField: 5
   <RequiredItem>k__BackingField: {fileID: 11400000, guid: d68f399aa0ba3124e95c8b1c863391fe, type: 2}
   <Level>k__BackingField: 2
-  <IsPassed>k__BackingField: 1
+  <IsPassed>k__BackingField: 0
   Dialogue:
-  - not passed
+  - dialogue
   PassDialogue:
   - you passed
   AlreadyPassedDialogue:
   - already passed
   NoPassDialogue:
-  - 4
+  - no pass
