+++ Assets/SchoolRPG/ScriptableObjects/NPC/6.asset
@@ -15,12 +15,12 @@ MonoBehaviour:
   <Name>k__BackingField: 6
   <RequiredItem>k__BackingField: {fileID: 11400000, guid: c5f760399dc615747b22875575454cdd, type: 2}
   <Level>k__BackingField: 3
-  <IsPassed>k__BackingField: 1
+  <IsPassed>k__BackingField: 0
   Dialogue:
-  - not passed
+  - dialogue
   PassDialogue:
   - you passed
   AlreadyPassedDialogue:
   - already passed
   NoPassDialogue:
-  - 4
+  - no pass
