+++ Assets/SchoolRPG/ScriptableObjects/NPC/7.asset
@@ -15,12 +15,12 @@ MonoBehaviour:
   <Name>k__BackingField: 7
   <RequiredItem>k__BackingField: {fileID: 11400000, guid: 5ea85c5dfcd044445a7c70b2bf9d29f0, type: 2}
   <Level>k__BackingField: 3
-  <IsPassed>k__BackingField: 1
+  <IsPassed>k__BackingField: 0
   Dialogue:
-  - not passed
+  - dialogue
   PassDialogue:
   - you passed
   AlreadyPassedDialogue:
   - already passed
   NoPassDialogue:
-  - 4
+  - no pass
