+++ Assets/SchoolRPG/Source/SchoolRPG/ProgressTracker/Runtime/TextFader.cs
@@ -9,19 +9,23 @@ public class TextFader : MonoBehaviour
     [SerializeField] public TextMeshProUGUI textDisplay;
     [SerializeField] private float fadeDuration = 1.0f;
 
-    public static TextFader Instance { get; private set; }
-
     public void Start()
     {
-        textDisplay.CrossFadeAlpha(0, 0.01f, true);
+        Color color = textDisplay.color;
+        color.a = 0;
+        textDisplay.color = color;
+        //textDisplay.CrossFadeAlpha(0, 0.01f, true);
+        textDisplay.enabled = false;
         // StartCoroutine(FadeInAndOut());
     }
 
     public IEnumerator FadeInAndOut()
     {
+        textDisplay.enabled = true;
         textDisplay.CrossFadeAlpha(1, fadeDuration, true);
         yield return new WaitForSeconds(2f);
         textDisplay.CrossFadeAlpha(0, fadeDuration, true);
+        textDisplay.enabled = false;
     }
 
     /*
