+++ Assets/SchoolRPG/Source/SchoolRPG/SceneManagement/Runtime/CanvasesDDOL.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 868acaad5dbbbfa4684036d56b3a69f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
