+++ Assets/SchoolRPG/Source/SchoolRPG/SceneManagement/Runtime/ScreenFader.cs
@@ -7,20 +7,7 @@ public class ScreenFader : MonoBehaviour
     [SerializeField] private Image fadeImage;
     [SerializeField] private float fadeDuration = 1.0f;
 
-    public static ScreenFader Instance { get; private set; }
-
-    private void Awake()
-    {
-        if (Instance == null)
-        {
-            Instance = this;
-            DontDestroyOnLoad(gameObject);
-        }
-        else
-        {
-            Destroy(gameObject);
-        }
-    }
+    
 
     private void Start()
     {
